add symlink checks

+++ condarecipe9.2/scripts/deactivate.d/cudatoolkit-dev-deactivate.sh
@@ -3,7 +3,10 @@
 for f in $CONDA_PREFIX/pkgs/cudatoolkit-dev-9.2-0/bin/*;
 do  
     to_unlink=$(basename ${f});
-    unlink $CONDA_PREFIX/bin/${to_unlink};
+    
+    if [-L $CONDA_PREFIX/bin/${to_unlink}]; then
+        unlink $CONDA_PREFIX/bin/${to_unlink};
+    fi 
 
 done
 
@@ -11,7 +14,10 @@ done
 for f in $CONDA_PREFIX/pkgs/cudatoolkit-dev-9.2-0/lib64/*;
 do  
     to_unlink=$(basename ${f});
-    unlink $CONDA_PREFIX/lib/${to_unlink};
+
+    if [-L $CONDA_PREFIX/lib/${to_unlink}]; then
+       unlink $CONDA_PREFIX/lib/${to_unlink};
+    fi 
     
 done
